Move `VFHSignature308` to feature types header
