+++ common/include/pcl/feature_types.h
@@ -96,6 +96,11 @@ namespace pcl
     * \ingroup common
     */
   struct FPFHSignature33;
+
+  /** \brief Members: float vfh[308]
+    * \ingroup common
+    */
+  struct VFHSignature308;
 } // namespace pcl
 /** @} */
