+++ common/src/point_types.cpp
@@ -233,14 +233,6 @@ namespace pcl
   }
 
   std::ostream& 
-  operator << (std::ostream& os, const VFHSignature308& p)
-  {
-    for (int i = 0; i < 308; ++i)
-    os << (i == 0 ? "(" : "") << p.histogram[i] << (i < 307 ? ", " : ")");
-    return (os);
-  }
-
-  std::ostream& 
   operator << (std::ostream& os, const BRISKSignature512& p)
   {
     os << p.scale << " " << p.orientation << " ";
