+++ features/include/pcl/features/vfh.h
@@ -42,7 +42,7 @@
 
 #include <array>
 
-#include <pcl/point_types.h>
+#include <pcl/feature_types.h>
 #include <pcl/features/feature.h>
 
 namespace pcl
