+++ tools/vfh_estimation.cpp
@@ -37,6 +37,7 @@
  *
  */
 
+#include <pcl/point_types.h>
 #include <pcl/io/pcd_io.h>
 #include <pcl/features/vfh.h>
 #include <pcl/console/print.h>
